+++ spec/PimEnterprise/Bundle/VersioningBundle/Denormalizer/Flat/ProductValue/AttributeOptionDenormalizerSpec.php
@@ -5,7 +5,7 @@ namespace spec\PimEnterprise\Bundle\VersioningBundle\Denormalizer\Flat\ProductVa
 use PhpSpec\ObjectBehavior;
 use Pim\Bundle\CatalogBundle\Entity\Repository\AttributeOptionRepository;
 use Pim\Bundle\CatalogBundle\Entity\AttributeOption;
-use Pim\Bundle\CatalogBundle\Model\AbstractAttribute;
+use Pim\Bundle\CatalogBundle\Model\AttributeInterface;
 use Pim\Bundle\CatalogBundle\Model\ProductValueInterface;
 
 class AttributeOptionDenormalizerSpec extends ObjectBehavior
@@ -23,13 +23,13 @@ class AttributeOptionDenormalizerSpec extends ObjectBehavior
         $this->shouldBeAnInstanceOf('Symfony\Component\Serializer\Normalizer\DenormalizerInterface');
     }
 
-    function it_denormalizes_attribute_option($repository, ProductValueInterface $productValueInterface, AbstractAttribute $abstractAttribute, AttributeOption $red)
+    function it_denormalizes_attribute_option($repository, ProductValueInterface $productValueInterface, AttributeInterface $color, AttributeOption $red)
     {
         $data = '1';
         $context['value'] = $productValueInterface;
 
-        $productValueInterface->getAttribute()->shouldBeCalled()->willReturn($abstractAttribute);
-        $abstractAttribute->getCode()->willReturn('color');
+        $productValueInterface->getAttribute()->shouldBeCalled()->willReturn($color);
+        $color->getCode()->willReturn('color');
 
         $repository->findByReference('color.1')->shouldBeCalled()->willReturn($red);