+++ spec/PimEnterprise/Bundle/WorkflowBundle/Comparator/MetricComparatorSpec.php
@@ -9,7 +9,7 @@ class MetricComparatorSpec extends ObjectBehavior
 {
     function let(
         Model\AbstractProductValue $value,
-        Model\AbstractAttribute $attribute
+        Model\AttributeInterface $attribute
     ) {
         $value->getAttribute()->willReturn($attribute);
     }