+++ spec/PimEnterprise/Bundle/WorkflowBundle/Presenter/AttributePresenterSpec.php
@@ -3,7 +3,7 @@
 namespace spec\PimEnterprise\Bundle\WorkflowBundle\Presenter;
 
 use PhpSpec\ObjectBehavior;
-use Pim\Bundle\CatalogBundle\Model\AbstractAttribute;
+use Pim\Bundle\CatalogBundle\Model\AttributeInterface;
 use Pim\Bundle\EnrichBundle\Twig\LocaleExtension;
 use Prophecy\Argument;
 
@@ -26,43 +26,43 @@ class AttributePresenterSpec extends ObjectBehavior
         $this->shouldBeAnInstanceOf('PimEnterprise\Bundle\WorkflowBundle\Presenter\PresenterInterface');
     }
 
-    function it_supports_product_attribute(AbstractAttribute $attribute)
+    function it_supports_product_attribute(AttributeInterface $attribute)
     {
         $this->supports($attribute, [])->shouldBe(true);
     }
 
-    function it_presents_unlocalizable_and_unscopable_attribute(AbstractAttribute $attribute)
+    function it_presents_unlocalizable_and_unscopable_attribute(AttributeInterface $attribute)
     {
         $attribute->isLocalizable()->willReturn(false);
         $attribute->isScopable()->willReturn(false);
-        $attribute->__toString()->willReturn('Name');
+        $attribute->getLabel()->willReturn('Name');
 
         $this->present($attribute, [])->shouldReturn('Name');
     }
 
-    function it_presents_localizable_but_unscopable_attribute(AbstractAttribute $attribute)
+    function it_presents_localizable_but_unscopable_attribute(AttributeInterface $attribute)
     {
         $attribute->isLocalizable()->willReturn(true);
         $attribute->isScopable()->willReturn(false);
-        $attribute->__toString()->willReturn('Name');
+        $attribute->getLabel()->willReturn('Name');
 
         $this->present($attribute, ['__context__' => ['locale' => 'en_US']])->shouldReturn('[en_US] - Name');
     }
 
-    function it_presents_unlocalizable_but_scopable_attribute(AbstractAttribute $attribute)
+    function it_presents_unlocalizable_but_scopable_attribute(AttributeInterface $attribute)
     {
         $attribute->isLocalizable()->willReturn(false);
         $attribute->isScopable()->willReturn(true);
-        $attribute->__toString()->willReturn('Name');
+        $attribute->getLabel()->willReturn('Name');
 
         $this->present($attribute, ['__context__' => ['scope' => 'ecommerce']])->shouldReturn('ecommerce - Name');
     }
 
-    function it_presents_localizable_and_scopable_attribute(AbstractAttribute $attribute)
+    function it_presents_localizable_and_scopable_attribute(AttributeInterface $attribute)
     {
         $attribute->isLocalizable()->willReturn(true);
         $attribute->isScopable()->willReturn(true);
-        $attribute->__toString()->willReturn('Name');
+        $attribute->getLabel()->willReturn('Name');
 
         $this
             ->present($attribute, [