+++ src/PimEnterprise/Bundle/WorkflowBundle/Presenter/AttributePresenter.php
@@ -11,7 +11,7 @@
 
 namespace PimEnterprise\Bundle\WorkflowBundle\Presenter;
 
-use Pim\Bundle\CatalogBundle\Model\AbstractAttribute;
+use Pim\Bundle\CatalogBundle\Model\AttributeInterface;
 
 /**
  * Present an attribute
@@ -27,7 +27,7 @@ class AttributePresenter implements PresenterInterface, TwigAwareInterface
      */
     public function supports($data, array $change)
     {
-        return $data instanceof AbstractAttribute;
+        return $data instanceof AttributeInterface;
     }
 
     /**
@@ -52,7 +52,7 @@ class AttributePresenter implements PresenterInterface, TwigAwareInterface
             $parts[] = $change['__context__']['scope'];
         }
 
-        $parts[] = (string) $data;
+        $parts[] = $data->getLabel();
 
         return join(' - ', $parts);
     }